redesign account tab media views

+++ app/src/main/java/com/owenlejeune/tvtime/api/tmdb/api/v3/model/AuthorDetails.kt
@@ -6,5 +6,5 @@ class AuthorDetails(
     @SerializedName("name") val name: String,
     @SerializedName("username") val username: String,
     @SerializedName("avatar_path") val avatarPath: String?,
-    @SerializedName("rating") val rating: Int
+    @SerializedName("rating") val rating: Float
 )
